# optimize multi_set for faster command and single node ring