--- a/Gemfile
+++ b/Gemfile
@@ -14,6 +14,7 @@ group :development, :test do
   gem 'rubocop-performance'
   gem 'rubocop-rake'
   gem 'simplecov'
+  gem 'stackprof'
 end
 
 group :test do