--- a/Gemfile
+++ b/Gemfile
@@ -14,7 +14,7 @@ group :development, :test do
   gem 'rubocop-performance'
   gem 'rubocop-rake'
   gem 'simplecov'
-  gem 'stackprof'
+  gem 'stackprof', platform: :mri
 end
 
 group :test do