--- a/bin/profile
+++ b/bin/profile
@@ -69,7 +69,7 @@ when 'get_multi'
 when 'get_single'
   profile = StackProf.run(mode: :wall, interval: 1_000) do
     5_000.times do
-      keys.keys.each do |key|
+      keys.each_key do |key|
         meta_client.get(key)
       end
     end