--- a/lib/dalli/client.rb
+++ b/lib/dalli/client.rb
@@ -96,6 +96,8 @@ module Dalli
     # Fetch multiple keys efficiently.
     # If a block is given, yields key/value pairs one at a time.
     # Otherwise returns a hash of { 'key' => 'value', 'key2' => 'value1' }
+    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Metrics/AbcSize
     def get_multi(*keys)
       keys.flatten!
       keys.compact!
@@ -104,12 +106,16 @@ module Dalli
 
       if block_given?
         pipelined_getter.process(keys) { |k, data| yield k, data.first }
+      elsif @options[:protocol]&.to_s == 'meta' && @ring.servers.size == 1
+        pipelined_getter.process(keys)
       else
         {}.tap do |hash|
           pipelined_getter.process(keys) { |k, data| hash[k] = data.first }
         end
       end
     end
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/AbcSize
 
     ##
     # Fetch multiple keys efficiently, including available metadata such as CAS.