… .github/workflows/rubocop.yml
… .github/workflows/tests.yml
… .rubocop.yml
… lib/dalli.rb
… lib/dalli/pipelined_setter.rb
… lib/dalli/protocol/connection_manager.rb
… lib/dalli/protocol/meta.rb
… lib/dalli/protocol/meta/response_processor.rb
… lib/dalli/protocol/value_compressor.rb
… test/helper.rb
… test/helpers/memcached.rb
… test/integration/test_operations.rb
… test/integration/test_pipelined_get.rb
… test/integration/test_pipelined_set.rb
… test/utils/memcached_manager.rb