We know that we were seeing some performance issues around multi-set, this PR sets up some benchmarking and profiling code and make some small improvements for multi_set.

The changes result in some small wins when benchmarked in just this repo:

```
Warming up --------------------------------------
write 100 keys simple
                        42.000 i/100ms
write_mutli 100 keys    49.000 i/100ms
Calculating -------------------------------------
write 100 keys simple
                        393.644 (± 5.8%) i/s    (2.54 ms/i) -      3.948k in  10.060753s
write_mutli 100 keys    449.569 (± 6.7%) i/s    (2.22 ms/i) -      4.508k in  10.071010s

Comparison:
write_mutli 100 keys:      449.6 i/s
write 100 keys simple:      393.6 i/s - 1.14x  slower
```

I also have a benchmark in a internal repo, showing it move from the implementation in that repo that used the `quiet` block to using the new multi_set and it also shows a nice reduction from be 3.5x slower to only 2x slower than our internal Rust memcached client. This internal benchmark is a slightly more real world test than the benchmarking code in the dalli repo.

benchmarks from internal performance checks using old loop with set in `quiet`:
```
set_multi 100 keys rust
                        99.000 i/100ms
set_multi 100 keys dalli
                        27.000 i/100ms
Calculating -------------------------------------
set_multi 100 keys rust
                        921.909 (± 5.6%) i/s    (1.08 ms/i) -      9.207k in  10.021302s
set_multi 100 keys dalli
                        258.824 (± 7.0%) i/s    (3.86 ms/i) -      2.592k in  10.063160s

Comparison:
set_multi 100 keys rust:      921.9 i/s
set_multi 100 keys dalli:      258.8 i/s - 3.56x  slower
```

new benchmarks using this set_multi method, significantly improves this going down to 2x difference

```
Warming up --------------------------------------
set_multi 100 keys rust
                        99.000 i/100ms
set_multi 100 keys dalli
                        46.000 i/100ms
Calculating -------------------------------------
set_multi 100 keys rust
                        928.040 (± 4.1%) i/s    (1.08 ms/i) -      9.306k in  10.045917s
set_multi 100 keys dalli
                        455.572 (± 7.5%) i/s    (2.20 ms/i) -      4.554k in  10.064782s

Comparison:
set_multi 100 keys rust:      928.0 i/s
set_multi 100 keys dalli:      455.6 i/s - 2.04x  slower
```